+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 `latest`
+- Added ChatUtil.SystemSendAllExcept
+- Added IsAdminCapable to Player
+- Added new ways to convert to Player model from Character Entity and User
+- New Vampires now add to PlayerService cache
+- Oakveil Woodlands added to World Regions
+- 1.0 VBlood Display Names now in Data.VBloods (1.1 VBloods will be added in a later update)
+
+`0.2.0`
 - Initial release